Update commercial license headers
Updated header.COMM to the files in tqtc-qtscxml. Examples, tests
or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals
with the earlier license header. Reason for this is
that some autotests use hard coded line numbers
and a change in the line count causes failures
in tests.

Task-number: QTQAINFRA-4220
Change-Id: I5705f5a36a3549b6db8aebe88149c2c4063dd2b4
Reviewed-by: Akseli Salovaara <akseli.salovaara@qt.io>