Fix calling of the overloaded Scxml statemachine methods
This commit reverts the commit
2951c124d3ff58abe1f84175e512ba26977f13d8 which marked the
dynamic statemachines' metabojects as fully dynamic. The marking
was done to prevent stale accesses via QML property cache.
However it introduced another problem with the resolution of the
QScxmlStateMachine's overloaded invokable methods. The result
was the inability to call all but one of the functions.

To address the original issues, instead of the dynamic marking,
this commit invalidates the related qml property cache during
the dynamic metaobject's destruction.

Task-number: QTBUG-89521
Task-number: QTBUG-92367
Change-Id: I5d206558bb052b27b2764a80c4bcbe6410c07b22
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>