Add missing find_package to an example
The find_package was not done for QOpenGLWidgets. This worked ok while
building as part of the repository because the package is sought at the
toplevel CMakeLists. However when building outside of the repo it is
necessary to have.

Task-number: QTBUG-92541
Change-Id: I3b69925e7c8b22ff44507ee6103f34e82787a227
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 9c159b6efad3fe12203fcebd693b63caec4caf1e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>