Update commercial license headers
Updated header.COMM to the files in tqtc-qtscxml.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4964
Change-Id: Ibc04a564e09e6b4264a86725bd54636f828e694d
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>