Add REUSE.toml files and missing licenses
REUSE.toml files are read by reuse to complement or override the
copyright and licensing information found in file.

The use of REUSE.toml files was introduced in REUSE version 3.1.0a1.
This reuse version is compatible with reuse specification
version 3.2 [1].
With this commit's files,
* The SPDX document generated by reuse spdx conforms to SPDX 2.3,
* The reuse lint command reports that the Qt project is reuse compliant.

In order to be reuse compliant all the licenses referenced in file
or within a REUSE.toml file must be present in the LICENSES
directory at the base of the module.
The missing licenses are added.

[1]: https://reuse.software/spec-3.2/

Task-number: QTBUG-124453
Task-number: QTBUG-125211
Change-Id: I00c00b2f4beaa0cd3e4d7837548921dacccc379a
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit dc629b20ac714173484ae29a566ee4745beef355)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>